refactor: rename base_loader to egg_loader
