+++ lib/egg_loader.js
